add user dtos

+++ src/main/java/azure/christmas_recipes/controllers/RecipeRestController.java
@@ -1,8 +1,11 @@
 package azure.christmas_recipes.controllers;
 
 import azure.christmas_recipes.exceptions.EmailAlreadyExistsException;
+import azure.christmas_recipes.exceptions.NoSuchUserException;
 import azure.christmas_recipes.models.dtos.ErrorDTO;
+import azure.christmas_recipes.models.dtos.RecipesDTO;
 import azure.christmas_recipes.models.dtos.UserDTO;
+import azure.christmas_recipes.models.dtos.UserRegistrationDTO;
 import azure.christmas_recipes.models.entities.User;
 import azure.christmas_recipes.services.IngredientService;
 import azure.christmas_recipes.services.RecipeIngredientsService;
@@ -12,6 +15,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import javax.persistence.criteria.CriteriaBuilder;
 import java.util.Optional;
 
 @RestController
@@ -28,15 +32,25 @@ public class RecipeRestController {
         this.userService = userService;
         this.recipeService = recipeService;
     }
+
+    @ExceptionHandler(NoSuchUserException.class)
+        public ResponseEntity handleNoSuchUser(NoSuchUserException e) {
+        return ResponseEntity.status(404).body(new ErrorDTO(e.getMessage()));
+    }
+
+    @GetMapping
+    public ResponseEntity ownProfile(@RequestHeader(value = "Authorizatioo") Integer id) {
+        Optional<UserDTO> user = userService.findById(id);
+        return ResponseEntity.status(200).body(user.get());
+    }
+
     @GetMapping("/{id}")
     public ResponseEntity findUserById(@PathVariable Integer id) {
-        Optional<UserDTO> user = userService.findById(id);
-        if (user.isPresent()) {
+            Optional<UserDTO> user = userService.findById(id);
             return ResponseEntity.status(200).body(user.get());
-        } else {
-            return ResponseEntity.status(404).body(new ErrorDTO("No such user"));
-        }
     }
+
+
     @PostMapping
     public ResponseEntity register(@RequestBody User user) {
         try {
@@ -44,16 +58,36 @@ public class RecipeRestController {
             return ResponseEntity.status(201).body(registeredUser) ;
         }catch (IllegalArgumentException e) {
             return ResponseEntity.status(406).body(new ErrorDTO(e.getMessage()));
-        } catch (EmailAlreadyExistsException e) {
-                return ResponseEntity.status(409).body(new ErrorDTO(e.getMessage()));
+        }
+    }
+
+
+    @PutMapping("/{id}")
+    public ResponseEntity modifyUser(@PathVariable Integer id, @RequestBody UserRegistrationDTO modification) {
+        try {
+            UserDTO modifiedUser = userService.update(id, modification);
+            return ResponseEntity.ok(modifiedUser);
+        } catch (IllegalArgumentException e) {
+            return ResponseEntity.status(406).body(new ErrorDTO(e.getMessage()));
+        }
+    }
+
+
+    @DeleteMapping("/{id}")
+    public ResponseEntity deleteUser(@PathVariable Integer id , @RequestHeader(value = "Authorization") Integer loggedInUser) {
+        if (id == loggedInUser) {
+            userService.delete(id);
+            return ResponseEntity.ok().build();
+        }else {
+            return ResponseEntity.status(403).body(new ErrorDTO("Don't have permission!"));
         }
 
     }
-    @DeleteMapping("/{id}")
-    public ResponseEntity deleteUser(@PathVariable Integer id) {
-        userService.delete(id);
-        return ResponseEntity.ok().build();
-    }
 
+    @GetMapping("/{id}/recipies")
+    public ResponseEntity getRecipiesOfUser(@PathVariable Integer id) {
+            RecipesDTO userFavouriteRecipies = recipeService.findUserRecipies(id);
+            return ResponseEntity.ok(userFavouriteRecipies);
+    }
 
 }
